--- modules/order/build.gradle.kts
@@ -1,37 +1,0 @@
-import org.springframework.boot.gradle.tasks.bundling.BootJar
-
-plugins {
-    alias(libs.plugins.kotlin.jpa)
-    alias(libs.plugins.kotlin.jvm)
-    alias(libs.plugins.kotlin.spring)
-    alias(libs.plugins.spring.boot)
-    alias(libs.plugins.spring.dependency.management)
-}
-
-java {
-    withSourcesJar()
-    withJavadocJar()
-}
-
-tasks.named<BootJar>("bootJar") {
-    enabled = false
-}
-
-tasks.named<Jar>("jar") {
-    enabled = true
-}
-
-dependencies {
-    implementation(project(":common:data"))
-    implementation(project(":common:exception"))
-    implementation(project(":common:identifier"))
-    implementation(project(":common:util"))
-
-    implementation(libs.kotlin.reflect)
-    implementation(libs.spring.boot.starter.data.jpa)
-    implementation(libs.spring.boot.starter.jooq)
-
-    runtimeOnly(libs.postgresql)
-
-    testImplementation(project(":common:test-support"))
-}
