+++ services/order-service/src/main/kotlin/com/nextmall/order/OrderApplication.kt
@@ -1,4 +1,4 @@
-package com.nextmall.orderservice
+package com.nextmall.order
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
@@ -6,18 +6,16 @@ import org.springframework.boot.runApplication
 
 @SpringBootApplication(
     scanBasePackages = [
-        "com.nextmall.orderservice",
         "com.nextmall.order",
         "com.nextmall.common",
     ],
 )
 @ConfigurationPropertiesScan(
-    "com.nextmall.orderservice",
     "com.nextmall.order",
     "com.nextmall.common",
 )
-class OrderServiceApplication
+class OrderApplication
 
 fun main(args: Array<String>) {
-    runApplication<OrderServiceApplication>(*args)
+    runApplication<OrderApplication>(*args)
 }
